New Repository
First running work-in-progress version of the electrical parameters datalogger based on "Raspberry PI Pico W" and on the energy meter Orno OR-WE-504.

The main script is auto-executable on startup saving it as "main.py" inside the Raspberry Pi Pico W board.
The script also require a "init.py" file where are stored all the main personal infos to establish the network connection and the mqtt connection.